created utils.py and moved utils functions
